requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2 